requirements.txt
langgraph
psycopg2-binary
langgraph-checkpoint-postgres>=2.0.0
psycopg-binary>=3.1.0
python-dotenv
fastapi
uvicorn
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
cryptography
langchain-google-genai
langchain-ollama
langchain-community
langchain-text-splitters
pypdf
python-multipart
